Address codex round 7 on #269: key per-generator state by identity, not array index
Rounds 2, 5, 6 and 7 all circled one root cause, so this fixes that instead of patching a fourth
symptom: the per-generator state was keyed by ARRAY INDEX, and an index is a position, not an
identity. Every removal shifted the rest onto each other's entries, which is why each round found
another way for a count or an in-flight flag to land on the wrong generator.

SenderRT now carries a `uid`, assigned once when a generator enters the runtime set, and both
gen_send_n and gen_firing are keyed by it. So:
- removing a generator no longer has to clear or re-key anything: the shift cannot hand one
  generator's count or in-flight flag to another, and a fire still in flight for the removed one
  writes back onto its own (now unused) entry;
- the epoch guard narrows to what it is actually for — the run/project boundary, where the
  sequences legitimately restart.

Also from this round:
- reset_gen_state() replaced the map and bumped the epoch without app.mu while a fire blocked in
  its transport send reads them under it (an unsafe concurrent map write). It takes the mutex now;
  its one caller runs before rebuild_from_proj acquires it, so there is no nesting.
- picking `counter` seeds `start` from the value the signal already sends, like the other sources
  (the seed chain had no counter branch, so it jumped to zero).

modules/ 76/76, saverule/loadrule green, GUI builds.

+++ cmd/blobly_net/run.v
@@ -677,7 +677,7 @@ fn (mut app App) start() {
 	// unsafe concurrent map write and the epoch a plain data race (codex #269).
 	app.mu.lock()
 	app.wave_t0_ns = time.sys_mono_now()
-	app.gen_send_n = map[int]int{} // a new run starts the per-send sequences at 0
+	app.gen_send_n = map[u64]int{} // a new run starts the per-send sequences at 0
 	app.gen_state_epoch++ // a fire still in flight from the previous run must not write into it
 	app.mu.unlock()
 	app.running = true
